+++ CHANGELOG.md
@@ -5,7 +5,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased][Unreleased]
 
+Added:
 
+1. (#429) Add *TelegramBot#getStickerSet* (by @CapacitorSet, @LibertyLocked)
 
 
 * * *